Shuffle around Distribution.Simple.Program.Types
This commmit removes the dependency of the
Distribution.Simple.Program.Types on Distribution.Simple.Program.Find
by shuffling around a few definitions between the modules.

This makes the "Types" module more self-contained, which means it
can be more easily imported without causing module cycles.
