+++ Cabal/src/Distribution/Simple/Program/Types.hs
@@ -24,7 +24,6 @@ module Distribution.Simple.Program.Types
     Program (..)
   , ProgramSearchPath
   , ProgramSearchPathEntry (..)
-  , simpleProgram
 
     -- * Configured program and related functions
   , ConfiguredProgram (..)
@@ -39,7 +38,6 @@ import Distribution.Compat.Prelude
 import Prelude ()
 
 import Distribution.PackageDescription
-import Distribution.Simple.Program.Find
 import Distribution.Verbosity
 import Distribution.Version
 
@@ -83,6 +81,29 @@ instance Show Program where
   show (Program name _ _ _ _) = "Program: " ++ name
 
 type ProgArg = String
+
+-- | A search path to use when locating executables. This is analogous
+-- to the unix @$PATH@ or win32 @%PATH%@ but with the ability to use
+-- the system default method for finding executables ('findExecutable' which
+-- on unix is simply looking on the @$PATH@ but on win32 is a bit more
+-- complicated).
+--
+-- The default to use is @[ProgSearchPathDefault]@ but you can add extra dirs
+-- either before, after or instead of the default, e.g. here we add an extra
+-- dir to search after the usual ones.
+--
+-- > ['ProgramSearchPathDefault', 'ProgramSearchPathDir' dir]
+type ProgramSearchPath = [ProgramSearchPathEntry]
+
+data ProgramSearchPathEntry
+  = -- | A specific dir
+    ProgramSearchPathDir FilePath
+  | -- | The system default
+    ProgramSearchPathDefault
+  deriving (Eq, Generic, Typeable)
+
+instance Binary ProgramSearchPathEntry
+instance Structured ProgramSearchPathEntry
 
 -- | Represents a program which has been configured and is thus ready to be run.
 --
@@ -144,22 +165,6 @@ programPath = locationPath . programLocation
 -- | Suppress any extra arguments added by the user.
 suppressOverrideArgs :: ConfiguredProgram -> ConfiguredProgram
 suppressOverrideArgs prog = prog{programOverrideArgs = []}
-
--- | Make a simple named program.
---
--- By default we'll just search for it in the path and not try to find the
--- version name. You can override these behaviours if necessary, eg:
---
--- > (simpleProgram "foo") { programFindLocation = ... , programFindVersion ... }
-simpleProgram :: String -> Program
-simpleProgram name =
-  Program
-    { programName = name
-    , programFindLocation = \v p -> findProgramOnSearchPath v p name
-    , programFindVersion = \_ _ -> return Nothing
-    , programPostConf = \_ p -> return p
-    , programNormaliseArgs = \_ _ -> id
-    }
 
 -- | Make a simple 'ConfiguredProgram'.
 --
